requirements.txt
openai
dotenv
python-docx